+++ scheduler/src/main/scala/mesosphere/marathon/core/instance/Reservation.scala
@@ -1,0 +1,8 @@
+package mesosphere.marathon
+package core.instance
+
+/**
+  * Represents a reservation for all resources that are needed for launching an instance
+  * and associated persistent local volumes.
+  */
+case class Reservation(volumeIds: Seq[LocalVolumeId], state: Reservation.State)
